move Net::LDAP::Connection to it's own file
